+++ orq_ai_sdk/client.py
@@ -1,8 +1,10 @@
-from typing import Optional
+from typing import Optional, Union
 
 from orq_ai_sdk.api_resources.async_deployments import Deployment as AsyncDeployment
+from orq_ai_sdk.api_resources.contacts.contacts import Contacts
 from orq_ai_sdk.api_resources.deployments import Deployment
 from orq_ai_sdk.api_resources.feedback.feedback import Feedback
+from orq_ai_sdk.api_resources.webhooks.webhooks import Webhooks
 from orq_ai_sdk.exceptions import OrqAIInvalidAPIException
 from orq_ai_sdk.models import Store
 
@@ -51,8 +53,34 @@ class OrqAI:
     def feedback(self):
         return Feedback()
 
+    @property
+    def webhooks(self):
+        return Webhooks()
+
+    @property
+    def contacts(self):
+        return Contacts()
+
     def set_user(self, id=None):
         Store["user_info"] = {"id": id}
+
+    def set_contact(
+            self,
+            id: str,
+            display_name: Union[str | None] = None,
+            email: Union[str | None] = None,
+            avatar_url: Union[str | None] = None,
+            metadata: Union[dict | None] = None
+    ):
+        Store['contact_id'] = id
+
+        self.contacts.create(
+            external_id=id,
+            display_name=display_name,
+            email=email,
+            avatar_url=avatar_url,
+            metadata=metadata
+        )
 
 
 class AsyncOrqAI:
@@ -99,5 +127,33 @@ class AsyncOrqAI:
     def feedback(self):
         return Feedback()
 
+    @property
+    def webhooks(self):
+        return Webhooks()
+
+    @property
+    def contacts(self):
+        return Contacts()
+
     def set_user(self, id=None):
         Store["user_info"] = {"id": id}
+
+    async def set_contact(
+            self,
+            id: str,
+            display_name: Union[str | None] = None,
+            email: Union[str | None] = None,
+            avatar_url: Union[str | None] = None,
+            metadata: Union[dict | None] = None
+    ):
+        Store['contact_id'] = id
+
+        await self.contacts.acreate(
+            external_id=id,
+            display_name=display_name,
+            email=email,
+            avatar_url=avatar_url,
+            metadata=metadata
+        )
+
+